swapped hw 3 and hw 4
