+++ go.mod
@@ -1,6 +1,6 @@
 module symbolic-execution-course
 
-go 1.23.2
+go 1.26.2
 
 require (
 	github.com/ebukreev/go-z3 v0.0.0-20250821144348-dfd1fde1462b
