+++ homework4/README.md
@@ -1,37 +1,47 @@
-# Домашнее задание 4: Внутрипроцедурный анализ основных конструкций
+# Домашнее задание 4: Символьная память
 
 ## Цель
-Реализовать основной цикл анализатора, а также интерпретацию основных языковых конструкций.
+Создать символьную память, которая в дальнейшем будет использоваться при символьной интерпретации программы.
 
 ## Задания
 
-### Задание 4.1: Реализация основного цикла анализатора
+### Задание 4.1: Реализация символьного выражения для ссылок
 
-В файле `analyser.go` необходимо реализовать функцию `Analyse(source string, functionName string) []Interpreter` таким образом,
-чтобы она выполняла основной цикл работы анализатора — доставала из `Analyzer.StatesQueue` при помощи `Analyser.PathSelector` следующее состояние
-интерпретатора для исполнения, вызывала шаг динамической интерпретации, и складывала все получившиеся состояния обратно в приоритетную очередь
-до тех пор, пока не сработает глобальная стратегия остановки (стратегию остановки выберите и реализуйте самостоятельно).
+Дополните свой набор символьных выражений символьным представлением ссылки. 
 
-### Задание 4.2*: Реализация Path Selectors
+### Задание 4.2: Реализация модели символьной памяти
 
-При желании, реализуйте дополнительные стратегии выбора пути в файле `path_selector.go`.
+Напишите модель символьной памяти, реализующую следующий интерфейс:
 
-### Задание 4.3: Интерпретация основных языковых конструкций
+```go
+type Memory interface {
+	Allocate(tpe symbolic.ExpressionType) *symbolic.Ref
 
-В файле `dynamic_interpreter.go` реализуйте следующие методы:
+	AssignField(ref *symbolic.Ref, fieldIdx int, value symbolic.SymbolicExpression)
 
-* `interpretDynamically(element ssa.Instruction) []Interpreter`, который по заданной ssa инструкции проводит шаг интерпретации
-и возвращает полученный набор новых состояний
-* `resolveExpression(value ssa.Value) symbolic.SymbolicExpression`, который по заданному expression получает символьное выражение.
+	GetFieldValue(ref *symbolic.Ref, fieldIdx int) symbolic.SymbolicExpression
 
-Интерпретация должна поддерживать все основные языковые конструкции - ветвления, примитивы, бинарные и унарные операции, циклы, ...
-За исключением вызовов функций, операций с массивами, объектами и указателями. На данном этапе требуется внутрипроцедурный анализ.
+	AssignToArray(ref *symbolic.Ref, index int, value symbolic.SymbolicExpression)
+
+	GetFromArray(ref *symbolic.Ref, index int) symbolic.SymbolicExpression
+}
+```
+
+Внутренне представление должно поддерживать алиасинг. 
+Интерфейс можно расширять и видоизменять для удобства дальнейшего использования при интерпретации (например, добавить функции `AllocateStruct`/`AllocateArray`).
+Рекомендуется реализовать модель символьной памяти из лекции (но это не обязательно).
+
+### Задание 4.3: Интеграция и тестирование
+
+Создайте тесты для демонстрации работы с символьной памятью, можно опираться на примеры, представленные в `homework3/examples/test_functions.go`.
+Для случаев алиасинга, в тестах должен вызываться SMT-солвер, кодирующий модель символьной памяти в SMT и демонстрирующий решение ограничений.
+Если ваша реализация структуры `memory.SymbolicMemory` основывается на символьных выражениях (использует именно символьные массивы, а не другие новые сущности),
+то для кодирования символьной памяти в смт запрос будет закономерно переиспользовать `translator.Z3Translator`.
 
 ## Что сдавать
 
 **Реализация в общей кодовой базе:**
-- `../internal/analyser.go`
-- `../internal/dynamic_interpreter.go`
-- `../internal/path_selector.go`
+- `../internal/symbolic/expressions.go` - реализуйте символьную ссылку
+- `../internal/memory/symbolic_memory.go` - реализуйте символьную память
