+++ homework4/examples/test_functions.go
@@ -1,207 +1,174 @@
 package main
 
-func test1(x int) int {
-	if x > 10 {
-		return x + 1
-	} else {
-		return x - 1
+type Person struct {
+	Name string
+	Age  int
+	ID   int
+}
+
+func testStructBasic() Person {
+	var p Person
+	p.Name = "Alice"
+	p.Age = 25
+	p.ID = 1001
+	return p
+}
+
+func testStructModification(p Person) Person {
+	p.Age = p.Age + 1
+	p.ID = p.ID * 2
+	return p
+}
+
+func testStructPointer() *Person {
+	p := &Person{Name: "Bob", Age: 30, ID: 2002}
+	p.Age = p.Age + 5
+	return p
+}
+
+func testStructPointerModification(p *Person) {
+	if p != nil {
+		p.Age = p.Age + 10
+		p.ID = p.ID + 1000
 	}
 }
 
-func test2(a, b bool) bool {
-	if a && b {
-		return true
-	}
-	return false
+func testNestedStructPointer() {
+	p := testStructPointer()
+	testStructPointerModification(p)
 }
 
-func testArithmetic(x, y int) int {
-	result := (x + y) * (x - y)
-
-	if result < 0 {
-		result = -result
+func testArrayFixed() [5]int {
+	var arr [5]int
+	for i := 0; i < 5; i++ {
+		arr[i] = i * i
 	}
-
-	return result % 10
+	return arr
 }
 
-func testUnary(x int) int {
-	y := -x
-	z := ^y
-
-	if z > 0 {
-		return z
+func testArrayModification(arr [5]int) [5]int {
+	for i := range arr {
+		arr[i] = arr[i] + 1
 	}
-	return y
+	return arr
 }
 
-func testComparisons(a, b int) bool {
-	if a == b {
-		return true
+func testSliceCreation() []int {
+	slice := make([]int, 5)
+	for i := range slice {
+		slice[i] = i * 2
 	}
-
-	if a != b && a > 0 {
-		return b < 0
-	}
-
-	return a >= b || a <= -b
+	return slice
 }
 
-func testLogicalOps(x, y int) bool {
-	cond1 := x > 0 && y > 0
-	cond2 := x < 0 || y < 0
-
-	return cond1 != cond2
+func testSliceAppend() []int {
+	var slice []int
+	for i := 0; i < 3; i++ {
+		slice = append(slice, i*10)
+	}
+	return slice
 }
 
-func testWhileLoop(n int) int {
-	i := 0
+func testSliceModification(slice []int) []int {
+	for i := range slice {
+		slice[i] = slice[i] * 2
+	}
+	slice = append(slice, 999)
+	return slice
+}
+
+type Student struct {
+	Name    string
+	Grades  [5]int
+	Average float64
+}
+
+func testStructWithArray() Student {
+	var s Student
+	s.Name = "Charlie"
+	s.Grades = [5]int{85, 90, 78, 92, 88}
+
 	sum := 0
-
-	for i < n {
-		sum += i
-		i++
+	for _, grade := range s.Grades {
+		sum += grade
 	}
-	return sum
+	s.Average = float64(sum) / float64(len(s.Grades))
+
+	return s
 }
 
-func testForLoop(n int) int {
-	result := 1
-	for i := 1; i <= n; i++ {
-		if i%2 == 0 {
-			result *= i
+type Address struct {
+	Street  string
+	City    string
+	ZipCode int
+}
+
+type Employee struct {
+	Person  Person
+	Address Address
+	Salary  float64
+}
+
+func testNestedStructs() Employee {
+	emp := Employee{
+		Person: Person{
+			Name: "David",
+			Age:  35,
+			ID:   3003,
+		},
+		Address: Address{
+			Street:  "Main St",
+			City:    "Boston",
+			ZipCode: 12345,
+		},
+		Salary: 75000.0,
+	}
+	return emp
+}
+
+func testNestedStructModification(emp *Employee) {
+	emp.Person.Age += 1
+	emp.Salary *= 1.1
+	emp.Address.ZipCode = 54321
+}
+
+func testArrayOfStructs() [3]Person {
+	var people [3]Person
+	people[0] = Person{Name: "Alice", Age: 25, ID: 1}
+	people[1] = Person{Name: "Bob", Age: 30, ID: 2}
+	people[2] = Person{Name: "Charlie", Age: 35, ID: 3}
+
+	people[1].Age += 5
+
+	return people
+}
+
+func testArrayOfStructsModification(people *[3]Person) {
+	for i := range people {
+		people[i].ID = people[i].ID * 10
+	}
+}
+
+type Foo struct {
+	a int
+}
+
+func Aliasing(foo1 *Foo, foo2 *Foo) int {
+	foo2.a = 5
+	foo1.a = 2
+	if foo2.a == 2 {
+		return 4
+	}
+	return 5
+}
+
+func testPathConstraintMutability(p Person) {
+	if p.Age != 18 {
+		p.Age = 18
+		if p.Age != 18 {
+			panic("Seems impossible")
 		} else {
-			result += i
+			println("Seems ok")
 		}
-	}
-	return result
-}
-
-func testInfiniteLoopBreak(x int) int {
-	i := 0
-	for {
-		if i >= x {
-			break
-		}
-		i++
-	}
-	return i
-}
-
-func testLoopWithConcreteBoundAndSymbolicBranching(condition bool) int {
-	result := 0
-	for i := 0; i < 10; i++ {
-		if condition && i%2 == 0 {
-			result += i
-		}
-	}
-	return result
-}
-
-func testLoopWithSymbolicBoundAndSymbolicBranching(n int, condition bool) int {
-	if n > 10 {
-		panic("Assumption violated: n should be less than or equal to 10")
-	}
-
-	result := 0
-	for i := 0; i < n; i++ {
-		if condition && i%2 == 0 {
-			result += i
-		}
-	}
-	return result
-}
-
-func testComplexConditions(a, b, c int) int {
-	if (a > b && b < c) || (a < b && b > c) {
-		return 1
-	} else if a == b && b == c {
-		return 2
-	} else {
-		return 3
-	}
-}
-
-func testNestedIf(x, y int) int {
-	if x > 0 {
-		if y > 0 {
-			return 1
-		} else {
-			return 2
-		}
-	} else {
-		if y > 0 {
-			return 3
-		} else {
-			return 4
-		}
-	}
-}
-
-func testBitwise(a, b int) int {
-	and := a & b
-	or := a | b
-	xor := a ^ b
-
-	if (and | or) > 0 {
-		return xor << 1
-	}
-	return xor >> 1
-}
-
-func testCombined(x int) int {
-	y := x * 2
-	z := 0
-
-	for i := 0; i < y; i++ {
-		if i%3 == 0 {
-			z += i
-		} else if i%3 == 1 {
-			z -= i
-		} else {
-			z *= 2
-		}
-
-		if z > 100 {
-			break
-		}
-	}
-
-	if z < 0 {
-		return -z
-	}
-	return z
-}
-
-func testEdgeCases(a int) int {
-	if a == 0 {
-		return 0
-	}
-
-	if a < 0 {
-		return -1
-	}
-
-	return 1
-}
-
-func testMultipleReturns(x int) int {
-	if x < 0 {
-		return -x
-	}
-
-	if x == 0 {
-		return 0
-	}
-
-	return x * 2
-}
-
-func testSimpleSum(a, b float64) float64 {
-	c := a + 1.1
-	if b+c > 10.1 && b+c < 11.125 {
-		return 1.1
-	} else {
-		return 1.2
 	}
 }
